Refactor code to clone all vms and fix issues with load balancing

+++ cslabs-backend/Proxmox/Responses/NodeStatus.cs
@@ -3,7 +3,7 @@ namespace CSLabsBackend.Proxmox.Responses
     public class NodeStatus
     {
         // 0- 100
-        public double CpuUsage { get; set; }
+        public int CpuUsage { get; set; }
         public MemoryUsage MemoryUsage { get; set; }
     }
 }
